+++ src/components/sidekick/chatbot.tsx
